+++ Runtime/SecureProtocol/SecureNetworkProtocol.cs
@@ -385,7 +385,7 @@ namespace Unity.Networking.Transport.TLS
         public NetworkProtocol CreateProtocolInterface()
         {
             return new NetworkProtocol(
-                computePacketAllocationSize: new TransportFunctionPointer<NetworkProtocol.ComputePacketAllocationSizeDelegate>(ComputePacketAllocationSize),
+                computePacketOverhead: new TransportFunctionPointer<NetworkProtocol.ComputePacketOverheadDelegate>(ComputePacketOverhead),
                 processReceive: new TransportFunctionPointer<NetworkProtocol.ProcessReceiveDelegate>(ProcessReceive),
                 processSend: new TransportFunctionPointer<NetworkProtocol.ProcessSendDelegate>(ProcessSend),
                 processSendConnectionAccept: new TransportFunctionPointer<NetworkProtocol.ProcessSendConnectionAcceptDelegate>(ProcessSendConnectionAccept),
@@ -402,10 +402,10 @@ namespace Unity.Networking.Transport.TLS
         }
 
         [BurstCompile(DisableDirectCall = true)]
-        [MonoPInvokeCallback(typeof(NetworkProtocol.ComputePacketAllocationSizeDelegate))]
-        public static int ComputePacketAllocationSize(ref NetworkDriver.Connection connection, IntPtr userData, ref int dataCapacity, out int dataOffset)
+        [MonoPInvokeCallback(typeof(NetworkProtocol.ComputePacketOverheadDelegate))]
+        public static int ComputePacketOverhead(ref NetworkDriver.Connection connection, out int dataOffset)
         {
-            return UnityTransportProtocol.ComputePacketAllocationSize(ref connection, userData, ref dataCapacity, out dataOffset);
+            return UnityTransportProtocol.ComputePacketOverhead(ref connection, out dataOffset);
         }
 
         public static bool ServerShouldStep(uint currentState)